Implement a bunch of comparison operators
I did not implement less than, greater than, less than or equal to; or
greater than or equal to with another register because it is ambiguous
how the logic should be done (signed vs unsigned). I did provide signed
and unsigned versions of these operators if compared against the inner
types themselves (like a > 0, etc).