Exposed the data stack inside the MultiInstructionExpression
Added a ton of documentation and a new variant function entry for
functions that can access the contents of the target Expression without
being a member of the class. Also, adding a the functions to the
expression via addInstruction no longer automatically causes them to be
deferred (this will allow future work with labels to become trivial).